+++ finagle-netty4/src/main/scala/com/twitter/finagle/netty4/transport/ChannelTransportContext.scala
@@ -1,13 +1,11 @@
 package com.twitter.finagle.netty4.transport
 
-import com.twitter.finagle.Status
 import com.twitter.finagle.transport.TransportContext
 import io.netty.channel.Channel
 import io.netty.handler.ssl.SslHandler
 import java.net.SocketAddress
 import java.security.cert.Certificate
 import java.util.concurrent.Executor
-import java.util.concurrent.atomic.AtomicBoolean
 import scala.util.control.NonFatal
 
 /**
@@ -17,13 +15,6 @@ import scala.util.control.NonFatal
 private[finagle] final class ChannelTransportContext(val ch: Channel)
     extends TransportContext
     with HasExecutor {
-
-  // Accessible by the `ChannelTransport` and for testing.
-  private[transport] val failed = new AtomicBoolean(false)
-
-  def status: Status =
-    if (failed.get || !ch.isOpen) Status.Closed
-    else Status.Open
 
   def localAddress: SocketAddress = ch.localAddress